Corrected type of minmapsize in table reading functions
The type of `minmapsize' was (strangely!) an `int' in the arguments of some
of the table reading functions. With this commit they are corrected to
`size_t'.

+++ lib/gnuastro/fits.h
@@ -257,7 +257,7 @@ gal_fits_tab_info(char *filename, char *hdu, size_t *numcols,
 gal_data_t *
 gal_fits_tab_read(char *filename, char *hdu, size_t numrows,
                   gal_data_t *colinfo, gal_list_sizet_t *indexll,
-                  int minmapsize);
+                  size_t minmapsize);
 
 void
 gal_fits_tab_write(gal_data_t *cols, gal_list_str_t *comments,
